more plugins

+++ Assets/Feel/NiceVibrations/Demo/Common/Scripts/NiceVibrationsDemoHelpers.cs
@@ -1,5 +1,7 @@
 // Copyright (c) Meta Platforms, Inc. and affiliates. 
 
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Lofelt.NiceVibrations
@@ -8,7 +10,7 @@ namespace Lofelt.NiceVibrations
     {
         public static float Round(float value, int digits)
         {
-            float mult = Mathf.Pow(10.0f, digits);
+            float mult = Mathf.Pow(10.0f, (float)digits);
             return Mathf.Round(value * mult) / mult;
         }
